fix: remove more files, refactor package name
