+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/data/model/merchant/GetMerchantByIdResponse.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.dash.wallet.features.exploredash.data.model.merchants
+package org.dash.wallet.features.exploredash.data.model.merchant
 
 import com.google.gson.annotations.SerializedName
